backend/pyproject.toml
[project]
name = "semantic-log-explorer-backend"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115",
    "uvicorn>=0.30",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "python-multipart>=0.0.9",
    "chromadb>=1.0",
    "google-genai>=1.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "hypothesis>=6.100",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v"
asyncio_mode = "auto"
markers = [
    "unit: Testes unitários",
    "property: Testes baseados em propriedades (Hypothesis)",
    "integration: Testes de integração",
]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "*/__pycache__/*", ".venv/*"]

[tool.coverage.report]
show_missing = true
fail_under = 80
exclude_lines = [
    "if __name__ == .__main__.",
    "pass",
    "\\.\\.\\.",
    "pragma: no cover",
]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "S",    # flake8-bandit (security)
    "T20",  # flake8-print
    "SIM",  # flake8-simplify
]
ignore = [
    "S101",  # assert allowed in tests
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101"]
